Refactor code for ground-truth
