+++ .gitignore
@@ -7,3 +7,9 @@ __pycache__
 .venv
 # wandb
 wandb
+# training
+checkpoints
+# data
+data
+# logs
+*.log
